Request movie from movieId clicked on item movie in recyclerView

+++ app/src/main/java/com/arctouch/codechallenge/api/TmdbApi.kt
@@ -33,7 +33,7 @@ interface TmdbApi {
 
     @GET("movie/{id}")
     fun movie(
-        @Path("id") id: Long,
+        @Path("id") id: Int,
         @Query("api_key") apiKey: String,
         @Query("language") language: String
     ): Observable<Movie>
